Update default config
Updated the default configuration for the application to include new prompts and other settings.
